moved superglue directly to repo
