+++ src/icepy4d/matching/match_pairs.py
@@ -8,8 +8,8 @@ from pathlib import Path
 
 from icepy4d.utils import AverageTimer
 
-from ..thirdparty.SuperGluePretrainedNetwork.models.matching import Matching
-from ..thirdparty.SuperGluePretrainedNetwork.models.utils import (
+from ..thirdparty.SuperGlue.models.matching import Matching
+from ..thirdparty.SuperGlue.models.utils import (
     make_matching_plot,
     frame2tensor,
     process_resize,
