+++ src/icepy4d/matching/superglue_tracker.py
@@ -28,10 +28,10 @@ from icepy4d.matching.utils import read_image, frame2tensor
 from icepy4d.utils import AverageTimer
 
 
-from icepy4d.thirdparty.SuperGluePretrainedNetwork.models.superpoint import SuperPoint
-from icepy4d.thirdparty.SuperGluePretrainedNetwork.models.superglue import SuperGlue
-from icepy4d.thirdparty.SuperGluePretrainedNetwork.models.matching import Matching
-from icepy4d.thirdparty.SuperGluePretrainedNetwork.models.utils import (
+from icepy4d.thirdparty.SuperGlue.models.superpoint import SuperPoint
+from icepy4d.thirdparty.SuperGlue.models.superglue import SuperGlue
+from icepy4d.thirdparty.SuperGlue.models.matching import Matching
+from icepy4d.thirdparty.SuperGlue.models.utils import (
     make_matching_plot,
 )
